requirements.txt
Flask>=2.0.0
flask-sqlalchemy
psycopg2-binary
python-dotenv
gunicorn